extract markdown rendering into seperate component
closes #204
